Add command line arguments to soccer_hardware.py
Also tried to make some stuff more OOP-like, but that wasn't the focus of this work session.

By default now, this script runs exactly as it would on the Jetson. However, command-line arguments are available that can change the serial port, the baud rate, the initial trajectory, and whether or not to use the ROS-related things. This way, the exact same script can be used by both the embedded and software teams, and all that changes is the parameters we run the script with. This ensures that developments on this interface remain consistent between teams