Fix re-init errors. Seems to work smoothly now, but some edge cases still need to be tested
Based on my testing, the MCU can now be disconnected and connected again smoothly—the automatic reocvery here seems to be working after this change. I also tested the case where the script is restarted while the microcontroller continues running, and the one time I tested that it worked as well.

TODO:
- Test trajectory loading
- Test in ROS environment
- Test timeouts to make sure they're working properly