Minor change to printing logic. RX and TX timeouts seem to be working
In the communicate method, I can comment out the line which transmits, or the line which receives, and in both cases the program still flows smoothly. This means the timeouts are working.

Additionally, I noticed that my PC was capable of about 250 transmissions per second when there was no receive logic in its way. This means we should easily be able to achieve a reliable 100 Hz refresh rate