First integration attempt
Transmission and reception are working, but need some work. Right now the transmission is occurring about 3x faster than the reception, which suggests that we might need to control the rate

Also, Ctrl+C is really not working nicely. When I kill it from a terminal, it hangs most of the time. Need to find a way to clean up all the threads and release the serial resources when this signal is sent.