Cleaned up Ctrl+C behaviour a bit, and worked on improving the timeout flexibility
We will still get serial exceptions upon Ctrl+C now because the serial resource will be released from main while Tx and Rx are still using it. The exception handling for this scenario seems to be working, but out of the 15 to 20 times I tried Ctrl+C, one of them still caused the program to hang when exiting.

Reliability this has been the main issue I've been trying to iron out. Our receive reliability seems pretty bad in the tests I've been doing, but I am using a L432KC to echo, and the program on it is not very smart, so that could be the cause of the issues. Will have to try with a F4 during the week